requirement.txt
certifi==2019.3.9
cffi==1.12.3
Click==7.0
cycler==0.10.0
easydict==1.9
fire==0.1.3
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
mkl-fft==1.0.12
mkl-random==1.0.2
numpy==1.16.4
olefile==0.46
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.2.1
Shapely==1.6.4.post2
six==1.12.0
torch==0.4.0
torchvision==0.2.1
tqdm==4.32.1
